refactor: changed App.jsx for react-router, moved old logic into TeamBuilder.jsx
